Separate port interfaces for Twitter API
