+++ src/twitter-adapter-playwright/UserProfileViewAdapter.ts
@@ -1,6 +1,6 @@
 import assert from "assert";
 import { Page } from "playwright";
-import { UserProfileView } from "../ObtainingUserDetails";
+import { UserProfileView } from "../twitter-api/ports/UserProfileView";
 
 export class UserProfileViewAdapter implements UserProfileView {
   private page: Page;
