+++ src/twitter-api/ports/UserProfileView.ts
@@ -1,0 +1,5 @@
+export type UserProfileView = {
+  open: (username: string) => Promise<void>;
+  readName: () => Promise<string>;
+  readBio: () => Promise<string | undefined>;
+};
